Refactored push and pull commands.
