+++ commands/commands_test.go
@@ -1,16 +1,5 @@
 package commands
 
-import (
-	"flag"
-	"io/ioutil"
-	"net/http"
-	"net/http/httptest"
-	"os"
-	"testing"
-
-	"github.com/urfave/cli"
-)
-
 const (
 	dasboardID         = "666"
 	apiKey             = "1234"
@@ -64,397 +53,3 @@ SRE - Kubernetes | renaud.hager@spam.com | hfy-m49-ps3| | 2019-04-03T18:27:49.04
 SRE - Vault Overview | renaud.hager@spam.com | hyv-dzi-xas| created by renaud.hager@spam.com (cloned)| 2019-02-27T12:03:28.403848+00:00| 2019-06-19T23:39:20.305904+00:00
 `
 )
-
-// TestGetDashboard test function and expect an error
-func TestGetDashboardReturnError(t *testing.T) {
-
-	_, _, err := getDashboard("https://myddenpoint", dasboardID, apiKey, appKey)
-
-	// We should have an error
-	if err == nil {
-		t.Errorf("getDashboard() didn't return an error")
-	}
-}
-
-// TestGetDashboardWrongResponseStatus test function to ensure that getDashboard() handle correctly return code
-func TestGetDashboardWrongResponseStatus(t *testing.T) {
-
-	ts := httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		w.WriteHeader(http.StatusServiceUnavailable)
-	}))
-	defer ts.Close()
-
-	_, statusCode, err := getDashboard(ts.URL, dasboardID, apiKey, appKey)
-
-	if err != nil {
-		t.Errorf("getDashboard() should not have returned an error")
-	}
-
-	if statusCode != 503 {
-		t.Errorf("getDashboard() should have 503 code")
-	}
-}
-
-// TestGetDashboardAssertRequest test function to ensure that getDashboard() send the correct request
-func TestGetDashboardAssertRequest(t *testing.T) {
-
-	ts := httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		w.WriteHeader(http.StatusOK)
-
-		if r.Method != "GET" {
-			t.Errorf("Expected 'GET' request, got '%s'", r.Method)
-		}
-
-		if r.RequestURI != "/api/v1/dashboard/"+dasboardID+"?api_key="+apiKey+"&application_key="+appKey {
-			t.Errorf("Did not get expected uri, got '%s'", r.RequestURI)
-		}
-
-		w.Write([]byte(datadogSuccessfullGetDashboard))
-
-	}))
-
-	defer ts.Close()
-
-	payload, _, err := getDashboard(ts.URL, dasboardID, apiKey, appKey)
-
-	if err != nil {
-		t.Errorf("getDashboard() should not have returned an error")
-	}
-
-	if payload != expectedPayloadGetDashboard {
-		t.Errorf("getDashboard() did not returned the expected payload")
-	}
-}
-
-// TestDumpDashboard test function for dumpDashboard()
-func TestDumpDashboard(t *testing.T) {
-
-	expectedContentByte := []byte(expectedContent)
-	notExistingFilePath := "/notexistingpath/myfile.json"
-	existingFilePath := "/tmp/dogsitter_coverage_test_file1.txt"
-
-	err := dumpDashboard(expectedContentByte, notExistingFilePath)
-
-	if err == nil {
-		t.Errorf("dumpDashboard() didn't return an error")
-	}
-
-	err = dumpDashboard(expectedContentByte, existingFilePath)
-
-	if err != nil {
-		t.Errorf("dumpDashboard() returned an error")
-	}
-
-	info, err := os.Stat(existingFilePath)
-
-	if err == nil {
-		mode := info.Mode().Perm().String()
-
-		if mode != dumpFilePermission {
-			t.Errorf("File has been created with wrong permissions expected '%s' got '%s'.", dumpFilePermission, mode)
-		}
-
-		content, _ := ioutil.ReadFile(existingFilePath)
-
-		if string(content) != expectedContent {
-			t.Errorf("dumpDashboard() content is not as expected. Expected '%s' got '%s'", string(content), expectedContent)
-		}
-
-		// Cleaning up
-		err = os.Remove(existingFilePath)
-
-		if err != nil {
-			t.Errorf("Error while cleaning up %s", existingFilePath)
-		}
-
-	} else if os.IsNotExist(err) {
-		t.Errorf("File %s has not been created", existingFilePath)
-	} else {
-		t.Errorf("Unknown error while testing file %s", existingFilePath)
-	}
-}
-
-// TestDumpDashboard test function for beautify()
-func TestBeautify(t *testing.T) {
-
-	payload := `{"a":"b"}`
-
-	prettyPayload := beautify(payload)
-
-	if string(prettyPayload) != expectedPrettyJSON {
-		t.Errorf("beautify() return is not correct. Expected '%s', got '%s'", expectedPrettyJSON, string(prettyPayload))
-	}
-}
-
-// TestLoadDashboard test function for loadDashboard()
-func TestLoadDashboard(t *testing.T) {
-
-	notExistingFilePath := "/notexistingpath/myfile.json"
-	existingFilePath := "/tmp/dogsitter_coverage_test_file1.txt"
-
-	_, err := loadDashboard(notExistingFilePath)
-
-	if err == nil {
-		t.Errorf("loadDashboard() did not return an error")
-	}
-
-	_ = ioutil.WriteFile(existingFilePath, []byte(expectedContent), 0644)
-
-	content, err := loadDashboard(existingFilePath)
-
-	if err != nil {
-		t.Errorf("loadDashboard() did return an error while reading %s", existingFilePath)
-	}
-	if string(content) != expectedContent {
-		t.Errorf("loadDashboard() did return the expected content, expected '%s' got '%s'", content, expectedContent)
-	}
-
-	// Cleaning up
-	err = os.Remove(existingFilePath)
-
-	if err != nil {
-		t.Errorf("Error while cleaning up %s", existingFilePath)
-	}
-}
-
-// TestUploadDashboard test function for uploadDashboard() handle correctly return code.
-func TestUploadDashboardWrongResponseStatus(t *testing.T) {
-
-	ts := httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		w.WriteHeader(http.StatusServiceUnavailable)
-	}))
-	defer ts.Close()
-
-	err := uploadDashboard(ts.URL, []byte(expectedContent), apiKey, appKey)
-
-	if err == nil {
-		t.Errorf("uploadDashboard() should have returned an error")
-	}
-}
-
-// TestUploadDashboardAssertRequest test function to ensure that uploadDashboard() send the correct request
-func TestUploadDashboardAssertRequest(t *testing.T) {
-
-	// Test when response is succesfull
-	ts := httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-
-		if r.Method != "POST" {
-			t.Errorf("Expected 'POST' request, got '%s'", r.Method)
-		}
-
-		if r.RequestURI != "/api/v1/dashboard?api_key="+apiKey+"&application_key="+appKey {
-			t.Errorf("Did not get expected uri, got '%s'", r.RequestURI)
-		}
-
-		if r.Header.Get("Content-Type") != "application/json" {
-			t.Errorf("Did not get expected HEADER, got %s", r.Header)
-		}
-
-		body, _ := ioutil.ReadAll(r.Body)
-
-		if string(body) != dashboardPayload {
-			t.Errorf("Did not get expected body,expected '%s' got %s", expectedPrettyJSON, string(body))
-		}
-
-		w.Write([]byte(datadogSuccessfullResponse))
-
-	}))
-
-	defer ts.Close()
-
-	err := uploadDashboard(ts.URL, []byte(dashboardPayload), apiKey, appKey)
-
-	if err != nil {
-		t.Errorf("uploadDashboard() should not have returned an error: %v", err)
-	}
-
-	// Test when response is failling
-	ts = httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-
-		w.Write([]byte("badJson"))
-
-	}))
-
-	defer ts.Close()
-
-	err = uploadDashboard(ts.URL, []byte(dashboardPayload), apiKey, appKey)
-
-	if err == nil {
-		t.Errorf("uploadDashboard() should  have returned an error")
-	}
-}
-
-// TestGetDashboardInfo function to test getDashboardInfo()
-func TestGetDashboardInfo(t *testing.T) {
-
-	id, url, err := getDashboardInfo(datadogSuccessfullResponse)
-
-	if err != nil {
-		t.Errorf("err should be nil, got %v", err)
-	}
-
-	if id != "dnq-s5w-h5j" {
-		t.Errorf("Did not get expected id, got %s", id)
-	}
-
-	if url != "/dashboard/dnq-s5w-h5j/sre---consul-overview" {
-		t.Errorf("Did not get expected url, got %s", url)
-	}
-
-	id, url, err = getDashboardInfo("badJson")
-
-	if err == nil {
-		t.Errorf("err should not be nil")
-	}
-
-	if len(id) != 0 {
-		t.Errorf("id should be empty, got %s", id)
-	}
-
-	if len(url) != 0 {
-		t.Errorf("url should be empty, got %s", url)
-	}
-}
-
-// TestStripBadField function that test stripBadField()
-func TestStripBadField(t *testing.T) {
-	input := `{"a":"b","c":"d","e":"f"}`
-	expectedOutput := `{"a":"b","e":"f"}`
-
-	output, err := stripBadField([]byte(input), "c")
-
-	if err != nil {
-		t.Errorf("err should be nil, got %v", err)
-	}
-
-	if string(output) != expectedOutput {
-		t.Errorf("output should be `%v`, got `%v`", expectedOutput, string(output))
-	}
-
-	_, err = stripBadField([]byte("foo"), "c")
-
-	if err == nil {
-		t.Errorf("err should not be nil")
-	}
-}
-
-// TestPullOK test for pull() with proper config
-func TestPullOK(t *testing.T) {
-
-	ts := httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		w.WriteHeader(http.StatusOK)
-
-		w.Write([]byte(datadogSuccessfullGetDashboard))
-
-	}))
-
-	defer ts.Close()
-
-	set := flag.NewFlagSet("test", 0)
-	set.String("dh", ts.URL, "doc")
-	set.String("api-key", apiKey, "doc")
-	set.String("app-key", appKey, "doc")
-	set.String("id", dasboardID, "doc")
-	set.String("o", "/tmp/test", "doc")
-
-	context := cli.NewContext(nil, set, nil)
-
-	err := pull(context)
-
-	if err != nil {
-		t.Errorf("pull() should not have returned an error, %v", err)
-	}
-	// Cleaning up
-	err = os.Remove("/tmp/test")
-	if err != nil {
-		t.Errorf("Error while cleaning up %s", "/tmp/test")
-	}
-
-}
-
-// TestPullKO test for pull() with wrong config
-func TestPullKO(t *testing.T) {
-
-	set := flag.NewFlagSet("test", 0)
-	set.String("dh", "wronghost", "doc")
-	set.String("api-key", apiKey, "doc")
-	set.String("app-key", appKey, "doc")
-
-	context := cli.NewContext(nil, set, nil)
-
-	err := pull(context)
-
-	if err == nil {
-		t.Errorf("pull() should have returned an error")
-	}
-
-	ts := httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		w.WriteHeader(http.StatusInternalServerError)
-	}))
-
-	defer ts.Close()
-
-	set2 := flag.NewFlagSet("test2", 0)
-	set2.String("dh", ts.URL, "doc")
-	set2.String("api-key", apiKey, "doc")
-	set2.String("app-key", appKey, "doc")
-
-	context2 := cli.NewContext(nil, set, nil)
-
-	err = pull(context2)
-
-	if err == nil {
-		t.Errorf("pull() should have returned an error")
-	}
-}
-
-// TestPushOK test for pull() with proper config
-func TestPushOK(t *testing.T) {
-
-	_ = ioutil.WriteFile("/tmp/test-push", []byte(expectedContent), 0644)
-
-	ts := httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-
-		w.Write([]byte(datadogSuccessfullResponse))
-
-	}))
-
-	defer ts.Close()
-
-	set := flag.NewFlagSet("test", 0)
-	set.String("dh", ts.URL, "doc")
-	set.String("api-key", apiKey, "doc")
-	set.String("app-key", appKey, "doc")
-	set.String("f", "/tmp/test-push", "doc")
-
-	context := cli.NewContext(nil, set, nil)
-
-	err := push(context)
-
-	if err != nil {
-		t.Errorf("push() should not have returned an error")
-	}
-
-	// Cleaning up
-	err = os.Remove("/tmp/test-push")
-	if err != nil {
-		t.Errorf("Error while cleaning up %s", "/tmp/test-push")
-	}
-}
-
-// TestPushKO test for pull() with wrong config
-func TestPushKO(t *testing.T) {
-
-	set := flag.NewFlagSet("test", 0)
-	set.String("dh", "wronghost", "doc")
-	set.String("api-key", apiKey, "doc")
-	set.String("app-key", appKey, "doc")
-
-	context := cli.NewContext(nil, set, nil)
-
-	err := push(context)
-
-	if err == nil {
-		t.Errorf("push() should have returned an error")
-	}
-}
