+++ commands/pull.go
@@ -12,6 +12,23 @@ import (
 	"github.com/urfave/cli"
 )
 
+// Pull interface
+type Pull interface {
+	pull(c *cli.Context) (err error)
+	getDashboard(ddEndpoint string, dashboardID string, apiKey string, appKey string) (string, int, error)
+	beautify(payload string) []byte
+	dumpDashboard(content []byte, filepath string) error
+	stripBadField(payload []byte, pattern string) ([]byte, error)
+}
+
+// PullAction struct
+type PullAction struct{}
+
+// NewPullAction constructor for DeleteAction
+func NewPullAction() Pull {
+	return &PullAction{}
+}
+
 func init() {
 	log.SetFormatter(&log.TextFormatter{
 		FullTimestamp: true,
@@ -23,7 +40,7 @@ func init() {
 var PullCmd = cli.Command{
 	Name:   "pull",
 	Usage:  "Pull dashboard configuration from Datadog API",
-	Action: pull,
+	Action: actionPull,
 	Flags: []cli.Flag{
 		cli.StringFlag{
 			Name:  "id",
@@ -37,9 +54,14 @@ var PullCmd = cli.Command{
 	},
 }
 
-func pull(c *cli.Context) (err error) {
+// actionPull placeholder function
+func actionPull(c *cli.Context) (err error) {
+	return NewPullAction().pull(c)
+}
 
-	payload, statusCode, err := getDashboard(c.GlobalString("dh"), c.String("id"), c.GlobalString("api-key"), c.GlobalString("app-key"))
+func (pa *PullAction) pull(c *cli.Context) (err error) {
+
+	payload, statusCode, err := pa.getDashboard(c.GlobalString("dh"), c.String("id"), c.GlobalString("api-key"), c.GlobalString("app-key"))
 
 	if err != nil {
 		log.Error("Error when querying Datadog API.")
@@ -51,13 +73,13 @@ func pull(c *cli.Context) (err error) {
 		return errors.New("Return code is not 200")
 	}
 
-	content := beautify(payload)
+	content := pa.beautify(payload)
 
 	if c.String("o") == "stdout" {
 		fmt.Print(string(content))
 		fmt.Print("\n")
 	} else {
-		err = dumpDashboard(content, c.String("o"))
+		err = pa.dumpDashboard(content, c.String("o"))
 		if err != nil {
 			return errors.New("Cannot dump the dashboard")
 		}
@@ -67,7 +89,7 @@ func pull(c *cli.Context) (err error) {
 }
 
 // getDashboard get dashboard JSON payload from Datadog
-func getDashboard(ddEndpoint string, dashboardID string, apiKey string, appKey string) (string, int, error) {
+func (pa *PullAction) getDashboard(ddEndpoint string, dashboardID string, apiKey string, appKey string) (string, int, error) {
 
 	var (
 		body       []byte
@@ -103,13 +125,13 @@ func getDashboard(ddEndpoint string, dashboardID string, apiKey string, appKey s
 	}
 
 	// Call to a function to strip a field that breaks Datadog API when the JSON is imported
-	strippedPayload, _ := stripBadField(body, "author_name")
+	strippedPayload, _ := pa.stripBadField(body, "author_name")
 
 	return string(strippedPayload), statusCode, err
 }
 
 // beautify JSON payload
-func beautify(payload string) []byte {
+func (pa *PullAction) beautify(payload string) []byte {
 
 	var (
 		output     []byte
@@ -129,7 +151,7 @@ func beautify(payload string) []byte {
 	return output
 }
 
-func dumpDashboard(content []byte, filepath string) error {
+func (pa *PullAction) dumpDashboard(content []byte, filepath string) error {
 
 	err := ioutil.WriteFile(filepath, content, 0600)
 	if err != nil {
@@ -144,7 +166,7 @@ func dumpDashboard(content []byte, filepath string) error {
 
 // stripBadField function is required because Datadog is broken.
 // We need to remove field `author_name` from the payload.
-func stripBadField(payload []byte, pattern string) ([]byte, error) {
+func (pa *PullAction) stripBadField(payload []byte, pattern string) ([]byte, error) {
 
 	var strippedPayload []byte
